fix fib function

+++ doctest_example.py
@@ -51,7 +51,7 @@ def fib(n):
         raise OverflowError("n is too large")
     
     # Here is the magic formula:
-    return (PHI**n - (-1/PHI)**n)/math.sqrt(5)
+    return int((PHI**n - (-1/PHI)**n)/math.sqrt(5))
     
     
     # change the return statement to the line below
